not_vulnerable/requirements.txt

Flask==3.0.2
Flask-SQLAlchemy==3.1.1
email-validator==2.1.1
itsdangerous==2.1.2
python-dotenv==1.0.1